rainy_sdk::models

Module catalog

Stable, domain-oriented model modules.

The original v0.x API keeps its compatibility types in this module. New contract-facing types live in these smaller modules and are re-exported below so applications can migrate without losing the ergonomic root API. Model-catalog domain facade.

Re-exports

pub use super::CapabilityFlag;

pub use super::ModelArchitecture;

pub use super::ModelBillingClass;

pub use super::ModelCatalogItem;

pub use super::ModelDataPolicy;

pub use super::ModelDataPolicyStatus;

pub use super::ModelDynamicPricing;

pub use super::ModelDynamicPricingTier;

pub use super::ModelLaunch;

pub use super::ModelLaunchAction;

pub use super::ModelLaunchAnimation;

pub use super::ModelLaunchControl;

pub use super::ModelLaunchGradient;

pub use super::ModelLaunchPresentation;

pub use super::ModelLaunchPricing;

pub use super::ModelLaunchSelection;

pub use super::ModelLaunchUi;

pub use super::ModelLaunchUiVariant;

pub use super::ModelLaunchVariant;

pub use super::ModelLaunchVariantPresentation;

pub use super::ModelList;

pub use super::ModelListItem;

pub use super::ModelPricing;

pub use super::ModelPricingPromo;

pub use super::ModelSelectionCriteria;

pub use super::ModelServiceTierPricing;

pub use super::ModelTier;

pub use super::RainyCapabilities;

pub use super::RainyCapabilitiesV2;

pub use super::RainyMultimodalCapabilitiesV2;

pub use super::RainyParametersCapabilitiesV2;

pub use super::RainyReasoningCapabilitiesV2;

pub use super::ReasoningControls;

pub use super::ReasoningMode;

pub use super::ReasoningPreference;

pub use super::ReasoningProfile;

pub use super::ReasoningProvider;

pub use super::ReasoningToggle;

pub use super::ThinkingBudget;